Updating the basic user Information
