+++ webapp/.gitignore
@@ -60,4 +60,4 @@ Thumbs.db
 .gradle/
 build/
 bin/
-.vscode
+/.vscode
