+++ webapp/src/main/webapp/app/registration-module/basic-information/UserBasicInfo.css
@@ -1,0 +1,54 @@
+
+.wrapper {
+  padding: 2rem;
+  padding-left: 4rem;
+  width: 55rem;
+  height: 35rem;
+  background-color: rgb(241, 245, 250);
+}
+
+.helpColor {
+  color: rgb(110, 19, 236);
+}
+
+.choicecards {
+  width: 50%; 
+  padding-left: 1rem; 
+  margin-bottom: 2rem;
+  padding-top: .5rem;
+  border-width: 1;
+  border-color: #E1E7ED90;
+  background-color: rgba(235, 241, 247, 0.712);
+}
+
+.iconStyle {
+  background-color: white;
+  padding-left: 1rem;
+  padding-right: 1rem;
+  padding-top: .4rem;
+  padding-bottom: .4rem;
+  border-radius: .5rem;
+  border: 2px solid #E1E7ED90;
+  margin-bottom: 1rem;
+  margin-top: 1rem;
+}
+
+.icon {
+  position: relative;
+  /* Adjust these values accordingly */
+  top: 2px;
+}
+
+.nextBtn {
+  border-width: 0;
+  border-radius: .4rem;
+  padding-left: 1rem;
+  padding-right: 1rem;
+  background-color: rgb(110, 19, 236);
+  text-align: center;
+}
+
+.nextBtn p {
+  top: -5
+}
+
